requirements.txt
pydantic
flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
numpy==2.3.2
MetaTrader5==5.0.5200
pandas==2.3.2
pydantic
flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
numpy==2.3.2
MetaTrader5==5.0.5200
pandas==2.3.2
werkzeug==3.1.3
python-dotenv
tzlocal
fastapi
urllib3
APScheduler
requests
pytz
google-auth
PyJWT